Creating single models.py file
